Split js into component files
